# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Supply Chain Proof-of-Delivery Escrow API
  version: 0.1.0
  description: >
    REST API for managing escrowed payments with proof-of-delivery attestation on the Stellar Soroban network.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:3000
    description: Local development server
security: []

paths:
  /orders:
    post:
      summary: Create a new escrow order
      description: >
        Deploys a new Soroban escrow contract instance and locks the buyer's funds in escrow.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully and funds locked in escrow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Missing or invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server or contract deployment error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all orders
      description: Returns all orders tracked in the local database.
      operationId: getAllOrders
      responses:
        "200":
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request query or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      summary: Get order details and live on-chain state
      description: Retrieves local tracking information along with real-time state read from the Soroban escrow contract.
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order details with live on-chain state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithChainState"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error or chain query failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/attest:
    post:
      summary: Attest proof of delivery
      description: An authorized attestor confirms delivery of the goods. Transitions order to Attested once M-of-N threshold confirmations are met.
      operationId: attestOrder
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttestOrderRequest"
      responses:
        "200":
          description: Order successfully attested or confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request or no local signer available for the attestor address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Created state, deadline has passed, or duplicate attestation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/claim:
    post:
      summary: Seller claims escrowed funds
      description: Seller withdraws funds from the escrow contract after successful attestation.
      operationId: claimOrder
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Funds claimed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: No local signer available for the seller address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Attested state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/reclaim:
    post:
      summary: Buyer reclaims escrowed funds after deadline
      description: Buyer reclaims funds if the delivery deadline has expired without attestation.
      operationId: reclaimOrder
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Funds reclaimed successfully by buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Created state or deadline has not yet passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/cancel:
    post:
      summary: Mutually cancel order before attestation
      description: Buyer and seller agree to cancel the order before attestation, refunding the buyer immediately.
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order cancelled successfully and funds returned to buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Missing local keypairs for buyer or seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Created state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/dispute:
    post:
      summary: Raise dispute against an attested order
      description: Buyer challenges an attestation before the seller claims funds, transitioning the order into Disputed state.
      operationId: disputeOrder
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order successfully transitioned to Disputed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Missing local keypair for buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Attested state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/resolve:
    post:
      summary: Resolve contested attestation dispute
      description: Designated arbiter resolves a disputed order, deciding whether to release funds to the seller or refund the buyer.
      operationId: resolveDispute
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeRequest"
      responses:
        "200":
          description: Dispute resolved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request body or missing local keypair for arbiter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Order is not in Disputed state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction execution failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - sellerAddress
        - amountStroops
        - deadlineSeconds
      properties:
        sellerAddress:
          type: string
          description: Stellar public address (G...) of the seller
          example: "GC5H3W256B3QW4A44GAK36XN763K2KRN7O2OESN553TUXW7R6AKN4V6E"
        attestorAddress:
          type: string
          description: Single Stellar public address (G...) of the attestor (legacy / convenience)
          example: "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
        attestors:
          type: array
          items:
            type: string
          description: Designated list of M-of-N Stellar public addresses authorized to confirm delivery
          example:
            - "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
            - "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        threshold:
          type: integer
          description: Minimum number of distinct attestations required (1 <= threshold <= len(attestors))
          example: 2
        arbiterAddress:
          type: string
          description: Optional designated arbiter address (G...) for dispute resolution; defaults to deployer keypair
          example: "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        amountStroops:
          type: string
          description: Amount in stroops (1 XLM = 10,000,000 stroops) as a string
          example: "10000000"
        deadlineSeconds:
          type: string
          description: UNIX timestamp (seconds) after which funds can be reclaimed if unattested
          example: "1735689600"
        tokenContractId:
          type: string
          description: Optional Stellar contract address (C...) of the payment token; defaults to configured native XLM asset if omitted
          example: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"

    AttestOrderRequest:
      type: object
      properties:
        attestorAddress:
          type: string
          description: Specific Stellar public address (G...) of the attestor signing the confirmation
          example: "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"

    ResolveDisputeRequest:
      type: object
      required:
        - releaseToSeller
      properties:
        releaseToSeller:
          type: boolean
          description: True to release escrowed funds to seller (Claimed); False to refund buyer (Reclaimed)
          example: true

    OrderTxHashes:
      type: object
      required:
        - create
        - attest
        - claim
        - reclaim
        - cancel
      properties:
        create:
          type: string
          nullable: true
          description: Transaction hash of escrow contract creation
          example: "a1b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f90"
        attest:
          type: string
          nullable: true
          description: Transaction hash of attestation
          example: null
        claim:
          type: string
          nullable: true
          description: Transaction hash of seller claim
          example: null
        reclaim:
          type: string
          nullable: true
          description: Transaction hash of buyer reclaim
          example: null
        cancel:
          type: string
          nullable: true
          description: Transaction hash of mutual cancellation
          example: null
        dispute:
          type: string
          nullable: true
          description: Transaction hash of raising a dispute
          example: null
        resolve:
          type: string
          nullable: true
          description: Transaction hash of dispute resolution
          example: null

    Order:
      type: object
      required:
        - id
        - contractId
        - buyerAddress
        - sellerAddress
        - attestorAddress
        - attestors
        - threshold
        - confirmations
        - tokenContractId
        - amountStroops
        - deadline
        - status
        - lifecycle
        - txHashes
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Local database order identifier
          example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
        contractId:
          type: string
          description: Stellar contract address (C...) of the escrow instance
          example: "CA3D5KRYMCMUZGAPOETEZ2NZNOME4H664X4UIRGFW6UR47X7U6GFF6MS"
        buyerAddress:
          type: string
          description: Stellar public address (G...) of the buyer
          example: "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        sellerAddress:
          type: string
          description: Stellar public address (G...) of the seller
          example: "GC5H3W256B3QW4A44GAK36XN763K2KRN7O2OESN553TUXW7R6AKN4V6E"
        attestorAddress:
          type: string
          description: Stellar public address (G...) of the primary/first attestor
          example: "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
        attestors:
          type: array
          items:
            type: string
          description: Authorized list of M-of-N attestor addresses
          example:
            - "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
            - "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        threshold:
          type: integer
          description: Number of required confirmations for attestation
          example: 2
        confirmations:
          type: array
          items:
            type: string
          description: List of attestor addresses that have confirmed delivery
          example:
            - "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
        arbiterAddress:
          type: string
          nullable: true
          description: Stellar public address (G...) of the designated arbiter
          example: "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        tokenContractId:
          type: string
          description: Stellar contract address (C...) of the payment asset
          example: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
        amountStroops:
          type: string
          description: Payment amount in stroops
          example: "10000000"
        deadline:
          type: integer
          description: UNIX timestamp deadline in seconds
          example: 1735689600
        status:
          type: string
          enum:
            - Created
            - Attested
            - Disputed
            - Claimed
            - Reclaimed
            - Cancelled
          description: On-chain / DB lifecycle state
          example: "Created"
        lifecycle:
          type: string
          enum:
            - in-transit
            - deadline-passed
            - delivered/confirmed
            - disputed
            - claimed
            - reclaimed
            - cancelled
          description: High-level human-readable lifecycle label
          example: "in-transit"
        txHashes:
          $ref: "#/components/schemas/OrderTxHashes"
        createdAt:
          type: string
          format: date-time
          description: ISO timestamp of order creation
          example: "2026-09-07T12:00:00.000Z"

    OnChainState:
      type: object
      required:
        - seller
        - attestors
        - threshold
        - confirmations
        - amount
        - deadline
        - status
      properties:
        seller:
          type: string
          description: Stellar address of the seller recorded on-chain
          example: "GC5H3W256B3QW4A44GAK36XN763K2KRN7O2OESN553TUXW7R6AKN4V6E"
        attestors:
          type: array
          items:
            type: string
          description: List of authorized attestor addresses recorded on-chain
          example:
            - "GD6W556Z365UFX3E4K54KPNK3R257K4O53EESK5Q3X7W25N5RN7OESQI"
        threshold:
          type: integer
          description: Confirmation threshold recorded on-chain
          example: 1
        confirmations:
          type: array
          items:
            type: string
          description: List of confirmed attestor addresses recorded on-chain
          example: []
        arbiter:
          type: string
          description: Arbiter address recorded on-chain
          example: "GB6NVEN5HSUBKMYCE5ZOWSK5RPO5RDTBWTJHQ35YKVGFFL3U2NOSVNQI"
        amount:
          type: string
          description: Escrowed amount in stroops
          example: "10000000"
        deadline:
          type: string
          description: UNIX timestamp deadline in seconds
          example: "1735689600"
        status:
          type: string
          enum:
            - Created
            - Attested
            - Disputed
            - Claimed
            - Reclaimed
            - Cancelled
          description: Raw on-chain status
          example: "Created"

    OrderWithChainState:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required:
            - onChain
          properties:
            onChain:
              $ref: "#/components/schemas/OnChainState"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing the reason for failure
          example: "Order 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d not found"